Fix size errors.

+++ kernel/include/arch/x86/paging/ia32e.h
@@ -14,18 +14,18 @@
  */
 struct IA32E_PageMapLevel4_Entry
 {
-    uint32 present        :  1;
-    uint32 write          :  1;
-    uint32 user           :  1;
-    uint32 writeThrough   :  1;
-    uint32 cacheDisable   :  1;
-    uint32 accessed       :  1;
-    uint32 ignored1       :  1;
-    uint32 pageSize       :  1; // must be 0
-    uint32 ignored2       :  4;
-    uint32 address        : 40; // M-1:12 is physical address, 51:M must be 0
-    uint32 ignored3       : 11;
-    uint32 executeDisable :  1;
+    uint64 present        :  1;
+    uint64 write          :  1;
+    uint64 user           :  1;
+    uint64 writeThrough   :  1;
+    uint64 cacheDisable   :  1;
+    uint64 accessed       :  1;
+    uint64 ignored1       :  1;
+    uint64 pageSize       :  1; // must be 0
+    uint64 ignored2       :  4;
+    uint64 address        : 40; // M-1:12 is physical address, 51:M must be 0
+    uint64 ignored3       : 11;
+    uint64 executeDisable :  1;
 } __attribute__((packed));
 
 /*
@@ -35,20 +35,20 @@ struct IA32E_PageMapLevel4_Entry
 // maps 4KB
 struct IA32E_PageTable_4KB_Entry
 {
-    uint32 present        :  1;
-    uint32 write          :  1;
-    uint32 user           :  1;
-    uint32 writeThrough   :  1;
-    uint32 cacheDisable   :  1;
-    uint32 accessed       :  1;
-    uint32 dirty          :  1;
-    uint32 pat            :  1;
-    uint32 global         :  1;
-    uint32 ignored1       :  3;
-    uint32 address        : 40; // M-1:12 is physical address, 51:M must be 0
-    uint32 ignored2       :  7;
-    uint32 protectionKey  :  4;
-    uint32 executeDisable :  1;
+    uint64 present        :  1;
+    uint64 write          :  1;
+    uint64 user           :  1;
+    uint64 writeThrough   :  1;
+    uint64 cacheDisable   :  1;
+    uint64 accessed       :  1;
+    uint64 dirty          :  1;
+    uint64 pat            :  1;
+    uint64 global         :  1;
+    uint64 ignored1       :  3;
+    uint64 address        : 40; // M-1:12 is physical address, 51:M must be 0
+    uint64 ignored2       :  7;
+    uint64 protectionKey  :  4;
+    uint64 executeDisable :  1;
 } __attribute__((packed));
 
 // maps 2MB
@@ -60,18 +60,18 @@ struct IA32E_PageTable_4KB
 // maps 2MB
 struct IA32E_PageDirectory_4KB_Entry
 {
-    uint32 present        :  1;
-    uint32 write          :  1;
-    uint32 user           :  1;
-    uint32 writeThrough   :  1;
-    uint32 cacheDisable   :  1;
-    uint32 accessed       :  1;
-    uint32 ignored1       :  1;
-    uint32 pageSize       :  1; // must be 0
-    uint32 ignored2       :  4;
-    uint32 address        : 40; // M-1:12 is physical address, 51:M must be 0
-    uint32 ignored3       : 11;
-    uint32 executeDisable :  1;
+    uint64 present        :  1;
+    uint64 write          :  1;
+    uint64 user           :  1;
+    uint64 writeThrough   :  1;
+    uint64 cacheDisable   :  1;
+    uint64 accessed       :  1;
+    uint64 ignored1       :  1;
+    uint64 pageSize       :  1; // must be 0
+    uint64 ignored2       :  4;
+    uint64 address        : 40; // M-1:12 is physical address, 51:M must be 0
+    uint64 ignored3       : 11;
+    uint64 executeDisable :  1;
 } __attribute__((packed));
 
 // maps 1GB
@@ -84,18 +84,18 @@ struct IA32E_PageDirectory_4KB
 // maps 1GB
 struct IA32E_PageDirectoryPointerTable_4KB_Entry
 {
-    uint32 present        :  1;
-    uint32 write          :  1;
-    uint32 user           :  1;
-    uint32 writeThrough   :  1;
-    uint32 cacheDisable   :  1;
-    uint32 accessed       :  1;
-    uint32 ignored1       :  1;
-    uint32 pageSize       :  1; // must be 0
-    uint32 ignored2       :  4;
-    uint32 address        : 40; // M-1:12 is physical address, 51:M must be 0
-    uint32 ignored3       : 11;
-    uint32 executeDisable :  1;
+    uint64 present        :  1;
+    uint64 write          :  1;
+    uint64 user           :  1;
+    uint64 writeThrough   :  1;
+    uint64 cacheDisable   :  1;
+    uint64 accessed       :  1;
+    uint64 ignored1       :  1;
+    uint64 pageSize       :  1; // must be 0
+    uint64 ignored2       :  4;
+    uint64 address        : 40; // M-1:12 is physical address, 51:M must be 0
+    uint64 ignored3       : 11;
+    uint64 executeDisable :  1;
 } __attribute__((packed));
 
 // maps 512GB
@@ -119,22 +119,22 @@ struct IA32E_PageMapLevel4_4KB
 // maps 2MB
 struct IA32E_PageDirectory_2MB_Entry
 {
-    uint32 present        :  1;
-    uint32 write          :  1;
-    uint32 user           :  1;
-    uint32 writeThrough   :  1;
-    uint32 cacheDisable   :  1;
-    uint32 accessed       :  1;
-    uint32 dirty          :  1;
-    uint32 pageSize       :  1; // must be 1
-    uint32 global         :  1;
-    uint32 ignored1       :  3;
-    uint32 pat            :  1;
-    uint32 alwaysZero     :  8;
-    uint32 address        : 31; // M-1:21 is physical address, 51:M must be 0
-    uint32 ignored2       :  7;
-    uint32 protectionKey  :  4;
-    uint32 executeDisable :  1;
+    uint64 present        :  1;
+    uint64 write          :  1;
+    uint64 user           :  1;
+    uint64 writeThrough   :  1;
+    uint64 cacheDisable   :  1;
+    uint64 accessed       :  1;
+    uint64 dirty          :  1;
+    uint64 pageSize       :  1; // must be 1
+    uint64 global         :  1;
+    uint64 ignored1       :  3;
+    uint64 pat            :  1;
+    uint64 alwaysZero     :  8;
+    uint64 address        : 31; // M-1:21 is physical address, 51:M must be 0
+    uint64 ignored2       :  7;
+    uint64 protectionKey  :  4;
+    uint64 executeDisable :  1;
 } __attribute__((packed));
 
 // maps 1GB
@@ -146,18 +146,18 @@ struct IA32E_PageDirectory_2MB
 // maps 1GB
 struct IA32E_PageDirectoryPointerTable_2MB_Entry
 {
-    uint32 present        :  1;
-    uint32 write          :  1;
-    uint32 user           :  1;
-    uint32 writeThrough   :  1;
-    uint32 cacheDisable   :  1;
-    uint32 accessed       :  1;
-    uint32 ignored1       :  1;
-    uint32 pageSize       :  1; // must be 0
-    uint32 ignored2       :  4;
-    uint32 address        : 40; // M-1:12 is physical address, 51:M must be 0
-    uint32 ignored3       : 11;
-    uint32 executeDisable :  1;
+    uint64 present        :  1;
+    uint64 write          :  1;
+    uint64 user           :  1;
+    uint64 writeThrough   :  1;
+    uint64 cacheDisable   :  1;
+    uint64 accessed       :  1;
+    uint64 ignored1       :  1;
+    uint64 pageSize       :  1; // must be 0
+    uint64 ignored2       :  4;
+    uint64 address        : 40; // M-1:12 is physical address, 51:M must be 0
+    uint64 ignored3       : 11;
+    uint64 executeDisable :  1;
 } __attribute__((packed));
 
 // maps 512GB
@@ -181,22 +181,22 @@ struct IA32E_PageMapLevel4_2MB
 // maps 1GB
 struct IA32E_PageDirectoryPointerTable_1GB_Entry
 {
-    uint32 present        :  1;
-    uint32 write          :  1;
-    uint32 user           :  1;
-    uint32 writeThrough   :  1;
-    uint32 cacheDisable   :  1;
-    uint32 accessed       :  1;
-    uint32 dirty          :  1;
-    uint32 pageSize       :  1; // must be 1
-    uint32 global         :  1;
-    uint32 ignored1       :  3;
-    uint32 pat            :  1;
-    uint32 alwaysZero     : 17;
-    uint32 address        : 22; // M-1:30 must be address, 51:M must be 0
-    uint32 ignored2       :  7;
-    uint32 protectionKey  :  4;
-    uint32 executeDisable :  1;
+    uint64 present        :  1;
+    uint64 write          :  1;
+    uint64 user           :  1;
+    uint64 writeThrough   :  1;
+    uint64 cacheDisable   :  1;
+    uint64 accessed       :  1;
+    uint64 dirty          :  1;
+    uint64 pageSize       :  1; // must be 1
+    uint64 global         :  1;
+    uint64 ignored1       :  3;
+    uint64 pat            :  1;
+    uint64 alwaysZero     : 17;
+    uint64 address        : 22; // M-1:30 must be address, 51:M must be 0
+    uint64 ignored2       :  7;
+    uint64 protectionKey  :  4;
+    uint64 executeDisable :  1;
 } __attribute__((packed));
 
 // maps 512GB
